FLOAT

+++ projecteDB.sql
@@ -53,8 +53,8 @@ CREATE TABLE EtakemonsDescription (
 CREATE TABLE EtakemonsPosition (
 	id INTEGER AUTO_INCREMENT NOT NULL,
 	idetakemon INTEGER NOT NULL,
-	lat DOUBLE,
-	lng DOUBLE,
+	lat FLOAT(10,5),
+	lng FLOAT(10,5),
 	tipoetakemon VARCHAR(40),
 	FOREIGN KEY (tipoetakemon) REFERENCES Etakemons(tipo) on delete cascade,
 	FOREIGN KEY (idetakemon) REFERENCES Etakemons(id) on delete cascade,
